refactor(sdk-go): split DID and VC methods
Move DID initialization and VC helper methods out of agent.go into agent_did.go as pure method moves.

Made-with: Cursor
